update project to 2023
