+++ .gitignore
@@ -4,6 +4,8 @@
 *.iws
 *.ipr
 
-# Gradle
-.gradle/
-build/
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
