+++ build.gradle
@@ -1,34 +1,28 @@
-buildscript {
-  repositories {
-    mavenCentral()
-    maven {
-      url 'http://oss.sonatype.org/content/repositories/snapshots'
-    }
-  }
-  dependencies {
-    classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:0.1-SNAPSHOT'
-  }
+plugins {
+    id 'org.jetbrains.kotlin.jvm' version '1.9.0'
+    id 'org.jetbrains.kotlin.plugin.serialization' version '1.9.0'
+    id 'io.ktor.plugin' version '2.3.4'
+    id 'application'
 }
 
-apply plugin: 'kotlin'
-apply plugin: 'application'
-
 repositories {
-  mavenCentral()
-  maven { url 'http://oss.sonatype.org/content/repositories/snapshots' }
-  maven { url 'http://repository.jetbrains.com/all' }
+    mavenCentral()
 }
 
 dependencies {
-  compile 'org.jetbrains.kotlin:kotlin-stdlib:0.1-SNAPSHOT'
-  compile 'org.wasabi:wasabi:0.1-SNAPSHOT'
-  compile 'org.json:json:20141113'
+    implementation("io.ktor:ktor-server-netty:2.3.4")
+    implementation 'org.json:json:20230227'
+    implementation 'org.slf4j:slf4j-api:1.7.36'
+    implementation 'org.slf4j:slf4j-simple:1.7.36'
 }
 
-mainClassName = 'poker.PlayerServiceKt'
-
-task wrapper(type: Wrapper) {
-  gradleVersion = '2.2'
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
 }
 
-task stage(dependsOn: ['installApp'])
+application {
+    mainClass = 'poker.player.kotlin.ApplicationKt'
+}
+
