+++ src/main/kotlin/poker/player/kotlin/Player.kt
@@ -1,4 +1,4 @@
-package poker
+package poker.player.kotlin
 
 import org.json.JSONObject
